Changes to locks and atomic types for correct concurrency
Bufferpool:
Every time `read_page_map` is used, it is actually done for the sake of getting
a read or write lock on the mapped frame. I replaced `read_page_map`
with `get_frame_by_ppid`. This should make it impossible for the frame
to be removed from the pool in the time between reading the map and
getting an `Arc`.

Table:
Improve the use of Atomic types in `insert()`. next_RID is atomically
incremented before anything else.